feat(component): add community header component
