+++ src/components/community/index.ts
@@ -1,4 +1,5 @@
 export * from './Post';
-export * from './Header';
+export * from './PostHeader';
 export * from './Chat';
 export * from './UserImage';
+export * from './Header';
