+++ src/screens/community/visibleType/index.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+import { CommunityHeader } from 'src/components';
+
+export const VisibleTypeScreen: React.FC = () => {
+  return <CommunityHeader title="공개 범위" />;
+};
